refactor(p2): extract OfficePro into sibling page (stage 4 of 4)
Final stage. SettingExperimentalOfficeProPage + 3 private composables
(OfficeProjectEditorDialog, WatchDocDialog, OfficeProSwitchRow) +
OFFICE_PROJECT_IMPORT_MAX_BYTES const + 7 private top-level helpers
(parseProjectList, parseProjectSourceRefs, appendSourceRef,
importOfficeProjectSource (suspend), String.toOfficeProjectId,
String.toSafeWorkspaceFileName, formatOfficeProjectUpdatedAt) moved into
new sibling SettingExperimentalOfficeProPage.kt.

Two non-contiguous source ranges in parent (bff24fac):
- L148-791 (composables)
- L1084-1168 (const + 7 helpers, ends at EOF)
Concatenated with a single blank separator in the new file.

Visibility change required: `ExperimentBooleanPill` widened from private
to internal. It lives in the coordinator (shared primitive group) but was
also called by OfficeProPage — cross-file extraction made the call site
private-inaccessible. This is the only allowed byte-equal exception
(modifier-only change).

Coordinator strip: 1168 -> 400 (-768 with import cleanup).
Imports trimmed: 36 this-refactor orphans.
Pre-existing dead imports left untouched (verified against cdbdf159):
base64Encode, FeishuOfficeAnalysisTemplate, CardGroup, navigateToChatPage,
LocalContext.

Baseline 6/463 unit tests preserved.
